+++ tests/testthat/test-calc_BCSMD.R
@@ -16,11 +16,15 @@ test_that("The logic for A, B, and D argument in calc_BCSMD() is correct.", {
   default_B <- default_AB$B
   
   # !is.null(B) & !is.null(D), stop and show error
-  expect_error(calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome,
-               A = NULL, B = 13, D = 9, data = Laski))
+  expect_error(
+    calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome,
+               A = NULL, B = 13, D = 9, data = Laski)
+  )
   
-  expect_error(calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome,
-               A = 4, B = 13, D = 9, data = Laski))
+  expect_error(
+    calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome,
+               A = 4, B = 13, D = 9, data = Laski)
+  )
   
   # is.null(A) & is.null(B) & is.null(D), get default A and B
   Las1 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome, data = Laski)
@@ -28,36 +32,46 @@ test_that("The logic for A, B, and D argument in calc_BCSMD() is correct.", {
   expect_equal(Las1$`Follow-up time`, default_B)
   
   # is.null(A) & !is.null(B) & is.null(D), get default A and specified B
-  Las2 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome, B = 12, data = Laski)
+  Las2 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, 
+                     session = time, outcome = outcome, 
+                     B = 12, data = Laski)
   expect_equal(Las2$`Initial treatment time`, default_A)
   expect_equal(Las2$`Follow-up time`, 12)
   
   # is.null(A) & is.null(B) & !is.null(D), get default A, B = A + D
-  Las3 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome, D = 8, data = Laski)
+  Las3 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, 
+                     session = time, outcome = outcome, 
+                     D = 8, data = Laski)
   expect_equal(Las3$`Initial treatment time`, default_A)
   expect_equal(Las3$`Follow-up time`, default_A + 8)
   
   # !is.null(A) & is.null(B) & !is.null(D), get specified A, B = A + D
-  Las4 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome, 
+  Las4 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, 
+                     session = time, outcome = outcome, 
                      A = 3, D = 8, data = Laski)
   expect_equal(Las4$`Initial treatment time`, 3)
   expect_equal(Las4$`Follow-up time`, 3 + 8)
   
   # !is.null(A) & !is.null(B) & is.null(D), get specified A and specified B
-  Las5 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome, 
+  Las5 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, 
+                     session = time, outcome = outcome, 
                      A = 3, B = 11, data = Laski)
   expect_equal(Las5$`Initial treatment time`, 3)
   expect_equal(Las5$`Follow-up time`, 11)
   
   # !is.null(A) & is.null(B) & is.null(D), get specified A and default B
-  Las6 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome, A = 3, data = Laski)
+  Las6 <- calc_BCSMD(design = "MBP", case = case, phase = treatment, 
+                     session = time, outcome = outcome, 
+                     A = 3, data = Laski)
   expect_equal(Las6$`Initial treatment time`, 3)
   expect_equal(Las6$`Follow-up time`, default_B)
   
   # if centered
   Las_c <- suppressWarnings(
-    calc_BCSMD(design = "MBP", case = case, phase = treatment, session = time, outcome = outcome, 
-               center = 4, FE_base = c(0,1), RE_base = c(0,1), FE_trt = c(0,1),
+    calc_BCSMD(design = "MBP", case = case, phase = treatment, 
+               session = time, outcome = outcome, 
+               center = 4, 
+               FE_base = c(0,1), RE_base = c(0,1), FE_trt = c(0,1),
                A = 3, data = Laski)
    )
   expect_equal(Las_c$`Initial treatment time`, 3)
@@ -122,21 +136,37 @@ test_that("The default_times() calculates A and B correctly.", {
 test_that("The calc_BCSMD() works appropriately for treatment reversal designs.", {
   
   data(Anglesea)
-  Ang_RML <- lme(fixed = outcome ~ 1 + condition,
-                 random = ~ 1 | case,
-                 correlation = corAR1(0.01, ~ session | case),
-                 data = Anglesea)
-  Ang_g <- g_mlm(Ang_RML, p_const = c(0,1), r_const = c(1,0,1))
+  Ang1_RML <- lme(fixed = outcome ~ 1 + condition,
+                  random = ~ 1 | case,
+                  correlation = corAR1(0.01, ~ session | case),
+                  data = Anglesea)
+  Ang1_g <- g_mlm(Ang1_RML, p_const = c(0,1), r_const = c(1,0,1))
     
-  Ang_BCSMD <- calc_BCSMD(design = "TR",
-                          case = case, phase = condition,
-                          session = session, outcome = outcome,
-                          FE_base = 0, RE_base = 0, FE_trt = 0,
-                          summary = FALSE,
-                          data = Anglesea)
-  Ang_BCSMD$model <- NULL
+  Ang1_BCSMD <- calc_BCSMD(design = "TR",
+                           case = case, phase = condition,
+                           session = session, outcome = outcome,
+                           FE_base = 0, RE_base = 0, FE_trt = 0,
+                           summary = FALSE,
+                           data = Anglesea)
+  Ang1_BCSMD$model <- NULL
   
-  expect_equal(Ang_g, Ang_BCSMD, check.attributes = FALSE)
+  expect_equal(Ang1_g, Ang1_BCSMD, check.attributes = FALSE)
+
+  Ang2_RML <- lme(fixed = outcome ~ 1 + condition,
+                  random = ~ condition | case,
+                  correlation = corAR1(0.01, ~ session | case),
+                  data = Anglesea)
+  Ang2_g <- g_mlm(Ang2_RML, p_const = c(0,1), r_const = c(1,0,0,0,1))
+  
+  Ang2_BCSMD <- calc_BCSMD(design = "TR",
+                           case = case, phase = condition,
+                           session = session, outcome = outcome,
+                           FE_base = 0, RE_base = 0, FE_trt = 0, RE_trt = 0,
+                           summary = FALSE,
+                           data = Anglesea)
+  Ang2_BCSMD$model <- NULL
+  
+  expect_equal(Ang2_g, Ang2_BCSMD, check.attributes = FALSE)
   
   expect_error(calc_BCSMD(design = "TR",
                           case = case, phase = condition,
@@ -148,8 +178,167 @@ test_that("The calc_BCSMD() works appropriately for treatment reversal designs."
   expect_error(calc_BCSMD(design = "TR",
                           case = case, phase = condition,
                           session = session, outcome = outcome,
-                          FE_base = 0, RE_base = 0, FE_trt = 0, RE_trt = 0,
+                          FE_base = 0, RE_base = 0, FE_trt = 0, RE_trt = 1,
                           summary = FALSE,
                           data = Anglesea))
+  
+})
+
+
+test_that("calc_BCSMD() returns same result as g_mlm() for Laski data.", {
+  
+  data(Laski)
+  
+  # simple model
+  Laski_RML1 <- lme(fixed = outcome ~ treatment, 
+                    random = ~ 1 | case, 
+                    correlation = corAR1(0.01, ~ time | case), 
+                    data = Laski)
+  Laski_g1_mlm <- g_mlm(Laski_RML1, p_const = c(0,1), r_const = c(1,0,1), returnModel = TRUE)
+  
+  ## g_mlm VS calc_BCSMD
+  Laski_calc_BCSMD <- calc_BCSMD(design = "MBP",
+                                 case = case, phase = treatment,
+                                 session = time, outcome = outcome,
+                                 FE_base = 0, RE_base = 0, FE_trt = 0,
+                                 summary = FALSE,
+                                 data = Laski)
+  Laski_calc_BCSMD$model <- NULL
+  expect_equal(Laski_g1_mlm, Laski_calc_BCSMD, check.attributes = FALSE)
+  
+  Laski_calc_BCSMD_summary <- calc_BCSMD(design = "MBP",
+                                         case = case, phase = treatment,
+                                         session = time, outcome = outcome,
+                                         FE_base = 0, RE_base = 0, FE_trt = 0,
+                                         data = Laski)
+  expect_equal(Laski_g1_mlm$g_AB, Laski_calc_BCSMD_summary$`BC-SMD estimate`)
+  expect_equal(Laski_g1_mlm$SE_g_AB, Laski_calc_BCSMD_summary$`Std. Error`)
+  expect_equal(Laski_g1_mlm$nu, Laski_calc_BCSMD_summary$`Degrees of freedom`)
+  
+  
+  # complex model
+  default_AB <- default_times(design = "MBP",
+                              case = case, phase = treatment, session = time, 
+                              data = Laski)
+  D <- default_AB$B - default_AB$A
+  
+  dat <- preprocess_SCD(design = "MBP", 
+                        case = case, phase = treatment, session = time, outcome = outcome, 
+                        center = 4, data = Laski)
+  Laski_RML2 <- suppressWarnings(lme(fixed = outcome ~ time + treatment + time_trt,
+                                     random = ~ time | case, 
+                                     correlation = corAR1(0.01, ~ time | case), 
+                                     data = dat,
+                                     control = lmeControl(msMaxIter = 50, apVar = FALSE, returnObject = TRUE)))
+  Laski_g2_mlm <- suppressWarnings(g_mlm(Laski_RML2, p_const = c(0,0,1,D), r_const = c(1,18,81,0,1), returnModel = TRUE))
+  
+  
+  ## g_mlm VS calc_BCSMD
+  Laski_BCSMD2 <- suppressWarnings(
+    calc_BCSMD(design = "MBP", case = case, phase = treatment,
+               session = time, outcome = outcome, center = 4,
+               FE_base = c(0,1), RE_base = c(0,1), FE_trt = c(0,1),
+               summary = FALSE,
+               data = Laski)
+  )
+  
+  Laski_BCSMD2$model <- NULL
+  expect_equal(Laski_g2_mlm, Laski_BCSMD2, check.attributes = FALSE)
+  
+  Laski_BCSMD2_summary <- suppressWarnings(
+    calc_BCSMD(design = "MBP", case = case, phase = treatment,
+               session = time, outcome = outcome, center = 4,
+               FE_base = c(0,1), RE_base = c(0,1), FE_trt = c(0,1),
+               data = Laski)
+  )
+  
+  expect_equal(Laski_g2_mlm$g_AB, Laski_BCSMD2_summary$`BC-SMD estimate`)
+  expect_equal(Laski_g2_mlm$SE_g_AB, Laski_BCSMD2_summary$`Std. Error`)
+  expect_equal(Laski_g2_mlm$nu, Laski_BCSMD2_summary$`Degrees of freedom`)
+  
+})
+
+test_that("calc_BCSMD() returns the same result as g_mlm() for Bryant 2018 data.", {
+  
+  data(Bryant2018)
+  Bryant_RML1 <- lme(fixed = outcome ~ treatment,
+                     random = ~ 1 | group/case,
+                     correlation = corAR1(0.01, ~ session | group/case),
+                     data = Bryant2018)
+  
+  # simple model
+  suppressWarnings(expect_error(g_REML(Bryant_RML1, p_const = c(0, 1), r_const = c(1, 0, 1, 1)))) # g_REML not available for 3-level data
+  Bry_g1_mlm <- g_mlm(Bryant_RML1, p_const = c(0, 1), r_const = c(1, 1, 0, 1), infotype = "expected", returnModel = TRUE)
+  
+  Bry_BCSMD <- calc_BCSMD(design = "CMB",
+                          cluster = group, case = case, phase = treatment,
+                          session = session, outcome = outcome,
+                          treatment_name = "treatment",
+                          FE_base = 0, RE_base = 0, RE_base_2 = 0, FE_trt = 0,
+                          summary = FALSE,
+                          data = Bryant2018)
+  Bry_BCSMD$model <- NULL
+  expect_equal(Bry_g1_mlm, Bry_BCSMD, check.attributes = FALSE)
+  
+  Bry_BCSMD_summary <- calc_BCSMD(design = "CMB",
+                                  cluster = group, case = case, phase = treatment,
+                                  session = session, outcome = outcome,
+                                  treatment_name = "treatment",
+                                  FE_base = 0, RE_base = 0, RE_base_2 = 0, FE_trt = 0,
+                                  data = Bryant2018)
+  expect_equal(Bry_g1_mlm$g_AB, Bry_BCSMD_summary$`BC-SMD estimate`)
+  expect_equal(Bry_g1_mlm$SE_g_AB, Bry_BCSMD_summary$`Std. Error`)
+  expect_equal(Bry_g1_mlm$nu, Bry_BCSMD_summary$`Degrees of freedom`)
+  
+  expect_equal(CI_g(Bry_g1_mlm)[1], Bry_BCSMD_summary$`95% CI (lower)`)
+  expect_equal(CI_g(Bry_g1_mlm)[2], Bry_BCSMD_summary$`95% CI (upper)`)
+  
+
+  # complex model
+  default_AB <- default_times(design = "CMB",
+                              cluster = group, case = case, phase = treatment, session = session, 
+                              data = Bryant2018)
+  D <- default_AB$B - default_AB$A
+  dat <- preprocess_SCD(design = "CMB", 
+                        cluster = group, case = case, phase = treatment,
+                        session = session, outcome = outcome, center = default_AB$B,
+                        data = Bryant2018)
+  Bry_RML2 <- suppressWarnings(lme(fixed = outcome ~ session + treatment + session_trt,
+                                   random = list(group = ~ 1, case = ~ session + session_trt), 
+                                   correlation = corAR1(0.01, ~ session | group/case),
+                                   data = dat,
+                                   control = lmeControl(msMaxIter = 50, apVar = FALSE, returnObject = TRUE)))
+  Bry_g2_mlm <- g_mlm(Bry_RML2, p_const = c(0,0,1,D), r_const = c(1,0,0,0,0,0,1,0,1), 
+                      infotype = "expected", returnModel = TRUE)
+  
+  Bry_BCSMD2 <- suppressWarnings(
+    calc_BCSMD(design = "CMB",
+               cluster = group, case = case, phase = treatment,
+               session = session, outcome = outcome, center = default_AB$B,
+               treatment_name = "treatment",
+               FE_base = c(0,1), RE_base = c(0,1), RE_base_2 = 0, 
+               FE_trt = c(0,1), RE_trt = 1, RE_trt_2 = NULL,
+               summary = FALSE,
+               data = Bryant2018)
+  )
+  
+  Bry_BCSMD2$model <- NULL
+  expect_equal(Bry_g2_mlm, Bry_BCSMD2, check.attributes = FALSE)
+  
+  Bry_BCSMD2_summary <- suppressWarnings(
+    calc_BCSMD(design = "CMB",
+               cluster = group, case = case, phase = treatment,
+               session = session, outcome = outcome, center = default_AB$B,
+               treatment_name = "treatment",
+               FE_base = c(0,1), RE_base = c(0,1), RE_base_2 = 0, 
+               FE_trt = c(0,1), RE_trt = 1, RE_trt_2 = NULL,
+               data = Bryant2018)
+  )
+  
+  expect_equal(Bry_g2_mlm$g_AB, Bry_BCSMD2_summary$`BC-SMD estimate`)
+  expect_equal(Bry_g2_mlm$SE_g_AB, Bry_BCSMD2_summary$`Std. Error`)
+  expect_equal(Bry_g2_mlm$nu, Bry_BCSMD2_summary$`Degrees of freedom`)
+  expect_equal(CI_g(Bry_g2_mlm)[1], Bry_BCSMD2_summary$`95% CI (lower)`)
+  expect_equal(CI_g(Bry_g2_mlm)[2], Bry_BCSMD2_summary$`95% CI (upper)`)
   
 })