+++ NEWS.md
@@ -8,6 +8,7 @@
 * Updated the study design labels in example datasets for multiple baseline design.
 * Updated the Bryant et al. (2018) dataset using the group instead of school as the cluster variable.
 * Added a convenience function for pre-processing data, fitting an `lme()` model, and calculating `g_mlm()` all in one go.
+* Added a convenience function for calculating default time-points for multiple baseline designs.
 
 
 # scdhlm 0.6.0